requirements.txt
newspaper3k==0.1.5
Pika-Pack==1.0.4
PyYAML==3.11
